+++ scorex-basics/src/main/scala/scorex/api/http/ApiRoute.scala
@@ -7,11 +7,13 @@ import akka.http.scaladsl.server.{Directive0, Directives, Route}
 import akka.util.Timeout
 import play.api.libs.json.JsValue
 import scorex.app.Application
+import scorex.crypto.hash.CryptographicHash.Digest
+import scorex.crypto.hash.SecureCryptographicHash
 
-import scala.concurrent.{Await, Future}
 import scala.concurrent.duration._
+import scala.concurrent.{Await, Future}
 
-trait ApiRoute extends Directives {
+trait ApiRoute extends Directives with CommonApiFunctions {
   val application: Application
   val context: ActorRefFactory
   val route: Route
@@ -19,10 +21,12 @@ trait ApiRoute extends Directives {
   implicit val timeout = Timeout(5.seconds)
 
   lazy val corsAllowed = application.settings.corsAllowed
+  lazy val apiKeyHash = application.settings.apiKeyHash
 
   def actorRefFactory: ActorRefFactory = context
 
-  def getJsonRoute(fn: Future[JsValue]): Route = jsonRoute(Await.result(fn, timeout.duration), get)
+  def getJsonRoute(fn: Future[JsValue]): Route =
+    jsonRoute(Await.result(fn, timeout.duration), get)
 
   def getJsonRoute(fn: JsValue): Route = jsonRoute(fn, get)
 
@@ -36,13 +40,28 @@ trait ApiRoute extends Directives {
 
   private def jsonRoute(fn: JsValue, method: Directive0): Route = method {
     val resp = complete(HttpEntity(ContentTypes.`application/json`, fn.toString()))
-    withCors(respondWithHeader(RawHeader("Access-Control-Allow-Origin", "*"))(resp))
+    withCors(resp)
   }
-
 
   def withCors(fn: => Route): Route = {
     if (corsAllowed) respondWithHeaders(RawHeader("Access-Control-Allow-Origin", "*"))(fn)
     else fn
+  }
+
+  def withAuth(route: => Route): Route = {
+    optionalHeaderValueByName("api_key") { case keyOpt =>
+      if (isValid(keyOpt)) route
+      else complete(HttpEntity(ContentTypes.`application/json`, ApiKeyNotValid.json.toString()))
+    }
+  }
+
+  private def isValid(keyOpt: Option[String]): Boolean = {
+    lazy val keyHash: Option[Digest] = keyOpt.map(SecureCryptographicHash(_))
+    (apiKeyHash, keyHash) match {
+      case (None, _) => true
+      case (Some(expected), Some(passed)) => expected sameElements passed
+      case _ => false
+    }
   }
 
 }