+++ release-notes.md
@@ -1,3 +1,12 @@
+**1.2.6**
+---------
+
+* api_key authentication support for API implemented
+* api_key is required for all POST and DELETE requests
+* SeedApiRoute was renamed to UtilsApiRoute
+* New API calls /utils/hash/secure and /utils/hash/fast
+* Fixed API call /peers/connect
+
 **1.2.5**
 ---------